fix creation of opportunities; additional cleanup on move to storage aggregator
