+++ src/services/infrastructure/storage-aggregator-resolver/storage.aggregator.resolver.module.ts
@@ -1,6 +1,5 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { StorageBucket } from '@domain/storage/storage-bucket/storage.bucket.entity';
 import { EntityResolverModule } from '../entity-resolver/entity.resolver.module';
 import { StorageAggregator } from '@domain/storage/storage-aggregator/storage.aggregator.entity';
 import { StorageAggregatorResolverService } from './storage.aggregator.resolver.service';
@@ -8,7 +7,6 @@ import { StorageAggregatorResolverService } from './storage.aggregator.resolver.
 @Module({
   imports: [
     EntityResolverModule,
-    TypeOrmModule.forFeature([StorageBucket]),
     TypeOrmModule.forFeature([StorageAggregator]),
   ],
   providers: [StorageAggregatorResolverService],
